Optional: Pets level with owner (#206)
* Changed stat types to short

Changed stat types to short to prevent unnecessary casts.

* Replaced GetLivingOwner()

Replaced two implementations for GetNPCOwner() and GetPlayerOwner() with a simpler GetLivingOwner() implementation.

Added Bladeturn spell type to CheckDefensiveSpells().

* Removed superfluous AutoSetStats() call

* Added overridden SortSpells() and ScalePetSpell()

Moved BD Subpet specific code out of GamePet.

* Changed log to private

Protected leads to inheriting classes misreporting their calling object in the log.

* Overrided OnSkillTrained() to apply pet scaling

Since BD subpets may have their spell scaling capped by BD spec, we have to re-scale subpet spells when a BD trains a spell line.

* Improved AutoSetStats()

Previous AutoSetStats() implementation would only set stats if the current stats were < 1.  This meant that a GM command changing a mob's level wouldn't change it's stats.

AutoSetStats() now checks the DB entry and npcTemplate to generate correct stats when the mob's level changes on multiple calls.

* Set default level to 0

This allows AutoSetStats() to skip calculations for newly created mobs that don't have a level assigned yet.

* Optional: pets now level up with owner

Pets now track the values they need from the spell that summoned them to allow them to change levels after being summoned.

GetPetLevel() was removed from summon spell handlers and replaced with GamePet.SetPetLevel().

SortSpells overidden to handle scaling pet spells.

AutoSetStats() was being called multiple times when a pet was summoned, so it's now more efficient and should be called less often.

* Otional: Pets level with owners

Pets now level up players with their owners if the server property is set.

BD subpets that are set to cap theri pet scaling level by BD spec have their spells scaled when the spell line they're from is respecced.

* Fixed spellcasting interruptions, added stat scaling

Changed OnAttackedByEnemy() and MakeAttack() to properly interrupt spellcasting when hit and being hit, not only when a second spell is queued.

AutoSetStats() now scaled pet stats with the npcTemplate values like other pets, allowing individual necro pet stats to be adjusted.  Since older DB entries set most stats to 30 by default, I added a warning log entry to notify them to update their DBs.

* Changed stats from int to short

* Changed stats from int to short

* Added pet_levels_with_owner & pet_cap_bd_minion_spell_scaling_by_spec

Also changed autoset server properties to short to reduce the number of casts needed.

* Removed GetPetLevel()

* Removed GetPetLevel()

Hardcoded level 44 cap isn't necessary, as pet level is already capped by the value of the summoning spell, and Eve already has Value set to 44 for all FnF spells.

* Removed GetPetLevel()

Removed GetPetLevel() as it's been moved to GamePet.SetPetLevel().

Level increase shouldn't be hardcoded. Instead, minion summoning spells should have their Damage set to -82 and Value set to 41 to match Live.

* Removed GetPetLevel()

Replaced by GamePet.SetPetLevel().

Hardcoded increase in level should instead be handled in the DB by changing necro pet summoning spell Damage to -90 and Value to 45, except for the minor zombie servant spell which should have Value set to 2.

* Removed superfluous AutoSetStats() call

* Changed log from protected to private

Protected leads to inherited classes reporting the wrong class in the log.

* Changed log from protected to private

Protected leads to inherited classes reporting the wrong class in the log.

* Removed GetPetLevel()

Replaced by GamePet.SetPetLevel().

* Removed log entries used during testing

+++ GameServer/gameutils/INpcTemplate.cs
@@ -89,14 +89,14 @@ namespace DOL.GS
 		/// <summary>
 		/// Gets the template npc stats
 		///</summary>
-		int Strength { get; }
-		int Constitution { get; }
-		int Dexterity { get; }
-		int Quickness { get; }
-		int Piety { get; }
-		int Intelligence { get; }
-		int Empathy { get; }
-		int Charisma { get; }
+		short Strength { get; }
+		short Constitution { get; }
+		short Dexterity { get; }
+		short Quickness { get; }
+		short Piety { get; }
+		short Intelligence { get; }
+		short Empathy { get; }
+		short Charisma { get; }
 
 		/// <summary>
 		/// Gets the template npc aggro values
